requirements.txt
django
djangorestframework
django-filters
dj-rest-auth
allauth
corsheaders